==> code/02-clean/01-clean-crsp.ipynb <==
# Have to install janitor package to run. Not included in anaconda.
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import math as math
import janitor

exec(open("../header.py").read())

# Purpose

To replicate and then test out Boone and White (2015).

# Import raw data

sample = False
header = Header(sample)

raw_crsp = pd.read_csv(header.raw_root("crsp_data.csv"), parse_dates = ['date'])

print(f"Date range: {min(raw_crsp.date)} - {max(raw_crsp.date)}")

raw_crsp.head(2)

# Clean data

# Basic cleaning
# Filter start date to May 1995 even though start date we care about is 1996
#   in order to establish ranking that carries forward until May 1996
cleaned_crsp = raw_crsp\
    .clean_names()\
    .loc[lambda x:(x.date >= '1995-05-01')&(x.date <= '2006-12-31')]\
    .dropna(axis = 0, subset = ['bid', 'ask', 'vol', 'prc', 'shrout'])\
    .assign(prc = lambda x:np.abs(x.prc))\
    .assign(mktcap = lambda x: x.prc * x.shrout)\
    .assign(year = lambda x:x.date.dt.year,
            month = lambda x:x.date.dt.month,
            yrmo = lambda x: x.year*100 + x.month)

# Variables of interest
cleaned_crsp = cleaned_crsp\
    .drop(['siccd', 'permco'], axis = 1)\
    .assign(bid_ask_spread = lambda x:(x.ask - x.bid)/((x.ask + x.bid)/2),
            turnover = lambda x:x.vol/x.mktcap,
            dollar_vol = lambda x:x.vol * x.prc)

# Cusip cleanup
cleaned_crsp = cleaned_crsp\
    .assign(cusip = lambda x:x.ncusip.str.replace('.0','',regex = False))\
    .assign(cusip = lambda x:x.ncusip.str.slice(0,8))\
    .drop(['ncusip'], axis = 1)

# Rank market cap by year-month
# cleaned_crsp['rank'] = cleaned_crsp\
#     .groupby('yrmo', as_index = False)['mktcap']\
#     .transform(lambda x:x.rank(ascending = False))

cleaned_crsp.head(5)

cleaned_crsp.columns

# Data Integrity checks

print(f"Old shape: {raw_crsp.shape}")
print(f"New shape: {cleaned_crsp.shape}")

# Permno is a unique identifier
assert cleaned_crsp\
    .groupby(['permno', 'date'], as_index = False)\
    .agg(counts = ('ticker', 'count'))\
    .loc[lambda x:x.counts > 1]\
    .shape[0] == 0

# CUSIP-date is a unique identifier
assert cleaned_crsp\
    .groupby(['cusip', 'date'], as_index = False)\
    .agg(counts = ('permno','count'))\
    .loc[lambda x:x.counts > 1, 'counts']\
    .shape[0] == 0

# Save cleaned datsets

cleaned_crsp.to_csv(header.clean_root("crsp.csv"), index = False)
