==> code/02-clean/02-clean-russell.ipynb <==
# Have to install janitor package to run. Not included in anaconda.
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import math as math
import janitor

exec(open("../header.py").read())

# Purpose

To replicate and then test out Boone and White (2015).

# Import raw data

sample = False
header = Header(sample)

raw_russ = pd.read_csv(header.raw_root("russell_index_inclusion.csv"), parse_dates = ['Date'])

raw_russ.head(5)

# Clean data

# Make russell indicator flags
# Add yrmo
# Substring cusip to first 8 to match crsp

cleaned_russ = raw_russ\
    .clean_names()\
    .loc[lambda x:(x.date >= '1996-01-01')&(x.date <= '2006-12-31')]\
    .assign(russell1000 = lambda x:np.select(
        [x.russell1000 == 'Y',x.russell1000 == 'N'],
        [1,0],
        default = -1),
            russell2000 = lambda x:np.select(
        [x.russell2000 == 'Y',x.russell2000 == 'N'],
        [1,0],
        default = -1),
            r1000_wt = lambda x:x.r1000_wt.fillna(0),
            r2000_wt = lambda x:x.r2000_wt.fillna(0)
    )\
    .assign(year = lambda x:x.date.dt.year,
            month = lambda x:x.date.dt.month,
            yrmo = lambda x: x.year*100 + x.month)\
    .assign(cusip = lambda x:x.cusip.str.slice(0,8))

cleaned_russ.head(5)

# Data integrity checks

assert len(cleaned_russ.russell1000.unique()) == 2
assert len(cleaned_russ.russell2000.unique()) == 2

# Date-cusip is unique identifier
assert cleaned_russ\
    .groupby(['date', 'cusip'], as_index = False)\
    .agg(counts = ('ticker', 'count'))\
    .loc[lambda x:x.counts > 1]\
    .shape[0] == 0

print(f"Old shape: {raw_russ.shape}")
print(f"New shape: {cleaned_russ.shape}")

# Save cleaned datsets

cleaned_russ.to_csv(header.clean_root("russell.csv"), index = False)
