==> code/02-clean/03-join-crsp-russell.ipynb <==
# Have to install janitor package to run. Not included in anaconda.
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import math as math
import janitor

exec(open("../header.py").read())

# Import raw data

sample = False
header = Header(sample)

raw_crsp = pd.read_csv(header.clean_root("crsp.csv"), parse_dates = ['date'])
raw_rs = pd.read_csv(header.clean_root("russell.csv"), parse_dates = ['date'])

raw_crsp.head(2)

raw_rs.head(2)

# Problems with CUSIP
* Example: Citigroup
* Link: https://wrds-www.wharton.upenn.edu/pages/support/support-articles/mergent-fisd/linking-mergent-compustat-or-crsp/

print(f"Russell CUSIP for Citigroup: {raw_rs.loc[lambda x:(x.ticker == 'C'),'cusip'].unique()}")
print(f"CRSP CUSIP for Citigroup: {raw_crsp.loc[lambda x:(x.ticker == 'C'),'cusip'].unique()}")

raw_crsp.loc[lambda x:(x.ticker == 'C')&(x.date >= '1998-11-12')&(x.date <= '1998-12-8'),['date','cusip']]

raw_rs.loc[lambda x:(x.ticker == 'C')&(x.date >= '1998-08-01')&(x.date <= '1999-02-01'),['date','cusip']]

# Pre-merge cleaning

rs = raw_rs\
    .drop(['date', 'mktvalue', 'shares', 'year', 'month'], axis = 1)

crsp = raw_crsp

# Join on month and cleaned cusip

rs_companies = rs\
    .loc[:,['cusip']]\
    .drop_duplicates()

# Has to be left so can create rankings for companies that just entered the Russell indices.
# Or else if inner, drops the past market cap data for those companies
# So do the first join to filter down to companies that are in Russell at some point in time
crsp_join = crsp\
    .merge(rs_companies, how = "inner", on = ["cusip"])\
    .merge(rs, how = "left", on = ["yrmo", "cusip"])

crsp_join.head(3)

print(f"CRSP // Original: {crsp.shape[0]}, Merged: {crsp_join.shape[0]}, Ratio: {crsp_join.shape[0]/crsp.shape[0]:.2f}")
print(f"Russell // Original: {rs.shape[0]}")
print(f"Unique  companies // Pre CRSP: {len(crsp.cusip.unique())}, Russell: {len(rs.cusip.unique())}, Post: {len(crsp_join.cusip.unique())}")

Missing ~400/7775 companies that should be in CRSP.

# Data integrity

crsp_join.columns

# Check that cusip-date is still a unique identifier
assert crsp_join\
    .groupby(['date', 'cusip'], as_index = False)\
    .agg(counts = ('ticker_x', 'count'))\
    .loc[lambda x:x.counts > 1]\
    .shape[0] == 0

# Validate that cusip match is reasonable
# Even when tickers don't match, clearly seem like the same company based on company name
crsp_join\
    .loc[lambda x:~x.ticker_y.isna()]\
    .loc[lambda x:x.ticker_x != x.ticker_y, ['ticker_x', 'ticker_y', 'tsymbol', 'comnam', 'name']]\
    .drop_duplicates()\
    .head(10)

# No invalid market cap, bid-ask spread, volume, turnover
assert np.sum(crsp_join.mktcap.isna()) == 0
assert np.sum(crsp_join.bid_ask_spread.isna()) == 0
assert np.sum(crsp_join.turnover.isna()) == 0
assert np.sum(crsp_join.dollar_vol.isna()) == 0

# Save cleaned datsets

crsp_join.to_csv(header.clean_root("crsp_russ.csv"), index = False)
