==> code/02-clean/04-create-crsp-rankings.ipynb <==
# Have to install janitor package to run. Not included in anaconda.
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import math as math
import janitor

exec(open("../header.py").read())

# Import raw data

sample = False
header = Header(sample)

raw_merged_df = pd.read_csv(header.clean_root("crsp_russ.csv"), parse_dates = ['date'])

raw_merged_df.columns

raw_merged_df.date.min()

# Create rank

ranks_df = raw_merged_df\
    .loc[lambda x:x.month == 5]\
    .groupby(['permno', 'yrmo'], as_index = False)\
    .apply(lambda d:d.loc[lambda x:x.date == max(x.date)])\
    .reset_index(drop = True)\
    .loc[:,['yrmo', 'year', 'permno', 'russell1000', 'mktcap']]\
    .fillna({'russell1000':0, 'russell2000':0})

ranks_df['mkt_rank_desc'] = ranks_df\
    .groupby(['yrmo','year','russell1000'], as_index = False)['mktcap']\
    .transform(lambda x:x.rank(ascending = False))

ranks_df['mkt_rank_asc'] = ranks_df\
    .groupby(['yrmo','year','russell1000'], as_index = False)['mktcap']\
    .transform(lambda x:x.rank(ascending = True))

ranks_df = ranks_df.reset_index(drop = True)\
    .rename({'year':'russell_year'}, axis = 1)\
    .loc[:,['russell_year', 'permno', 'mkt_rank_desc', 'mkt_rank_asc']]

# Create russell year variable

merged_df = raw_merged_df\
    .assign(russell_year = lambda x:np.where(x.month <= 5, x.year - 1, x.year))\
    .fillna({'russell1000':0, 'russell2000':0})

# Join on rank

merged_df = merged_df\
    .merge(ranks_df, how = 'left', on = ['permno', 'russell_year'])

merged_df = merged_df\
    .loc[lambda x:(x.russell1000 == 1)|(x.russell2000 == 1)]

merged_df.columns

# Data Integrity

## Example of contradictory market caps and reasoning

**Question**: Why is the larger company in the Russell 2000 while the smaller company is in the Russell 1000?

**Answer**: It's because the Russell 1000 company's market cap dropped from last year May and/or the Russell 2000 company's market cap rose from last year May. Russell 1000/2000 membership is mostly decided at month-end in May. The smaller market-cap company then dropped to Russell 2000 for the next year.

merged_df.loc[lambda x:(x.cusip == '88160910')&(x.yrmo >= 200505),
              ['yrmo','cusip','mktcap','russell1000','russell2000','prc','shrout','ticker_x', 'ticker_y', 'comnam', 'name']]

merged_df.loc[lambda x:(x.cusip == '92552R40')&(x.yrmo >= 200504),
              ['yrmo','cusip','mktcap','russell1000','russell2000','prc','shrout','ticker_x', 'ticker_y', 'comnam', 'name']]

## Other checks

# Check ratio of companies that don't have rankings.
np.sum(merged_df['mkt_rank_asc'].isna())/len(merged_df)

# Check that most are from 1996 or because of IPOs after May that join Russell 1000/2000 straight out the gate
merged_df.loc[lambda x:x.mkt_rank_asc.isna(),'year'].value_counts()

merged_df.loc[lambda x:x.mkt_rank_asc.isna(),'month'].value_counts()

# Either russell 1000 or russell 2000. Not both
assert np.sum((merged_df['russell1000'] + merged_df['russell2000']) != 1) == 0

# Permno-date is unique identifier
assert merged_df\
    .groupby(['permno','date'], as_index = False)\
    .agg(counts = ('cusip', 'count'))\
    .reset_index(drop = True)\
    .loc[lambda x:x.counts > 1]\
    .shape[0] == 0

# Russell 1000 should have larger market cap on average
t = merged_df\
    .groupby('russell1000', as_index = False)\
    .agg(mean_mkt_cap = ('mktcap', 'mean'))

r1000 = t.loc[lambda x:x.russell1000 == 1,'mean_mkt_cap'].values[0]
r2000 = t.loc[lambda x:x.russell1000 == 0,'mean_mkt_cap'].values[0]

assert r1000 > r2000

# Final filters

merged_df.columns

print(f"Original: {raw_merged_df.shape[0]}, New: {merged_df.shape[0]}")

# Save cleaned datsets

merged_df.to_csv(header.clean_root("crsp_russ_ranks.csv"), index = False)

merged_df_sample = merged_df.sample(int(0.05 * merged_df.shape[0]))
merged_df_sample.to_csv(header.sample_clean_root("crsp_russ_ranks.csv"), index = False)
